推进调整提示词加载逻辑，进行部分重构
- 将CoreModel明确为“表达模块”，移动至modules包
- 将模块相关内容移动到Modules包
- 为存在的单例实现添加双重锁定
- 调整父类Model的提示词加载逻辑，并修改了执行`chat`操作的逻辑
- 将MemorySummarizer中的几个总结逻辑拆分成三个子模块，但入口仍为MemorySummarizer
- 在README添加了许可声明

+++ src/main/java/work/slhaf/agent/common/chat/ChatClient.java
@@ -22,8 +22,8 @@ public class ChatClient {
     private String apikey;
     private String model;
 
-    private int top_p;
-    private int temperature;
+    private double top_p;
+    private double temperature;
     private int max_tokens;
 
     public ChatClient(String url, String apikey, String model) {
