+++ src/main/java/work/slhaf/agent/common/chat/pojo/ChatBody.java
@@ -14,9 +14,9 @@ public class ChatBody {
     @NonNull
     private List<Message> messages;
     @Builder.Default
-    private int temperature = 1;
+    private double temperature = 1;
     @Builder.Default
-    private int top_p = 1;
+    private double top_p = 1;
     private boolean stream;
     @Builder.Default
     private int max_tokens = 1024;
